Initial Commit : Base Service and IOT applicaiton
This is the initial commit, the service works but the IOT does not. it's contents is not functional. the Service includes a method to install and uninstall but no other trama code that would be required of a service to safely end process.

NO Azure code is included with this commit.